requirements.txt
fastapi == 0.111.0
pydantic == 2.8.2
uvicorn == 0.30.1
python-dateutil == 2.9.0.post0